+++ packages/mds-stream-processor/connectors/index.ts
@@ -17,33 +17,4 @@
 import { StreamProducer, StreamConsumer } from '@mds-core/mds-stream/stream-interface'
 
 export type StreamSource<TMessage> = (processor: (message: TMessage) => Promise<void>) => StreamConsumer
-export type StreamTransform<TMessageIn, TMessageOut> = (message: TMessageIn) => Promise<TMessageOut | null>
 export type StreamSink<TMessage> = () => StreamProducer<TMessage>
-
-export interface StreamProcessorController {
-  start: () => Promise<void>
-  stop: () => Promise<void>
-}
-
-export const StreamProcessor = <TMessageIn, TMessageOut>(
-  source: StreamSource<TMessageIn>,
-  transform: StreamTransform<TMessageIn, TMessageOut>,
-  sink: StreamSink<TMessageOut>
-): StreamProcessorController => {
-  const producer = sink()
-  const consumer = source(async message => {
-    const transformed = await transform(message)
-    if (transformed) {
-      await producer.write(transformed)
-    }
-  })
-  return {
-    start: async () => {
-      await producer.initialize()
-      await consumer.initialize()
-    },
-    stop: async () => {
-      await Promise.all([consumer.shutdown(), producer.shutdown()])
-    }
-  }
-}
