+++ packages/mds-stream-processor/index.ts
@@ -13,4 +13,4 @@
     See the License for the specific language governing permissions and
     limitations under the License.
  */
-export * from './processors'
+export { VehicleEventProcessor } from './processors/vehicle-event-processor'
